New issue: Set ".tmgit" instead of ".dotfiles" as git dir
The `.dotfiles` dir was used as a test and ended up being used until today. So, it would be better to use `.tmgit` instead of a generic name like `.dotfiles`.

However, we should check if there is a recommend way to migrate all existing data: if by moving the directory or if by using `git init`.